Fix test failures on Perl 5.26 w/ PERL_USE_UNSAFE_INC=0
Perl 5.26+ removes '.' from the default @INC, and this breaks the
mechanism internally where @INC-prefixes are stripped from fid_fileinfo
names ( in ::FileInfo.pm ).

This breaks tests due to the sample 'rdt' files expecting single-token
comparators relative to 't/', instead getting a fully qualified path in
the "got" side of the test.

This remedies this by ensuring a fully-qualified path to "t/" is in
@INC for the aforementioned prefix-stripper to pick-up on.

Bug: https://github.com/timbunce/devel-nytprof/issues/108
Bug: https://bugs.gentoo.org/615734
Bug: https://github.com/timbunce/devel-nytprof/pull/112